+++ udata/core/dataset/api.py
@@ -67,7 +67,7 @@ from .exceptions import (
 from .rdf import dataset_to_rdf
 
 
-DEFAULT_SORTING = '-created_at'
+DEFAULT_SORTING = '-created_at_internal'
 SUGGEST_SORTING = '-metrics.followers'
 
 
@@ -75,7 +75,7 @@ class DatasetApiParser(ModelApiParser):
     sorts = {
         'title': 'title',
         'created': 'created_at_internal',
-        'last_modified': 'last_modified_internal',
+        'last_update': 'last_modified_internal',
         'reuses': 'metrics.reuses',
         'followers': 'metrics.followers',
         'views': 'metrics.views',
@@ -113,7 +113,8 @@ class DatasetApiParser(ModelApiParser):
         if args.get('granularity'):
             datasets = datasets.filter(spatial__granularity=args['granularity'])
         if args.get('temporal_coverage'):
-            datasets = datasets.filter(temporal_coverage__start__gte=args['temporal_coverage'][:9], temporal_coverage__start__lte=args['temporal_coverage'][11:])
+            datasets = datasets.filter(temporal_coverage__start__gte=args['temporal_coverage'][:9],
+                                       temporal_coverage__start__lte=args['temporal_coverage'][11:])
         if args.get('featured'):
             datasets = datasets.filter(featured=args['featured'])
         if args.get('organization'):
@@ -210,7 +211,7 @@ class DatasetAPI(API):
         if dataset.deleted and request_deleted is not None:
             api.abort(410, 'Dataset has been deleted')
         DatasetEditPermission(dataset).test()
-        dataset.last_modified_internal = datetime.now()
+        dataset.last_modified_internal = datetime.utcnow()
         form = api.validate(DatasetForm, dataset)
         return form.save()
 
@@ -222,8 +223,8 @@ class DatasetAPI(API):
         if dataset.deleted:
             api.abort(410, 'Dataset has been deleted')
         DatasetEditPermission(dataset).test()
-        dataset.deleted = datetime.now()
-        dataset.last_modified_internal = datetime.now()
+        dataset.deleted = datetime.utcnow()
+        dataset.last_modified_internal = datetime.utcnow()
         dataset.save()
         return '', 204
 
@@ -333,7 +334,7 @@ class ResourcesAPI(API):
             api.abort(400, 'This endpoint only supports remote resources')
         form.populate_obj(resource)
         dataset.add_resource(resource)
-        dataset.last_modified_internal = datetime.now()
+        dataset.last_modified_internal = datetime.utcnow()
         dataset.save()
         return resource, 201
 
@@ -357,12 +358,13 @@ class ResourcesAPI(API):
 class UploadMixin(object):
     def handle_upload(self, dataset):
         prefix = '/'.join((dataset.slug,
-                           datetime.now().strftime('%Y%m%d-%H%M%S')))
+                           datetime.utcnow().strftime('%Y%m%d-%H%M%S')))
         infos = handle_upload(storages.resources, prefix)
         if 'html' in infos['mime']:
             api.abort(415, 'Incorrect file content type: HTML')
         infos['title'] = os.path.basename(infos['filename'])
-        checksum_type = next(checksum_type for checksum_type in CHECKSUM_TYPES if checksum_type in infos)
+        checksum_type = next(checksum_type for checksum_type in CHECKSUM_TYPES
+                             if checksum_type in infos)
         infos['checksum'] = Checksum(type=checksum_type, value=infos.pop(checksum_type))
         infos['filesize'] = infos.pop('size')
         del infos['filename']
@@ -382,7 +384,7 @@ class UploadNewDatasetResource(UploadMixin, API):
         infos = self.handle_upload(dataset)
         resource = Resource(**infos)
         dataset.add_resource(resource)
-        dataset.last_modified_internal = datetime.now()
+        dataset.last_modified_internal = datetime.utcnow()
         dataset.save()
         return resource, 201
 
@@ -429,7 +431,7 @@ class UploadDatasetResource(ResourceMixin, UploadMixin, API):
         for k, v in infos.items():
             resource[k] = v
         dataset.update_resource(resource)
-        dataset.last_modified_internal = datetime.now()
+        dataset.last_modified_internal = datetime.utcnow()
         dataset.save()
         if fs_filename_to_remove is not None:
             storages.resources.delete(fs_filename_to_remove)
@@ -484,9 +486,9 @@ class ResourceAPI(ResourceMixin, API):
         # update_resource saves the updated resource dict to the database
         # the additional dataset.save is required as we update the last_modified date.
         form.populate_obj(resource)
-        resource.last_modified_internal = datetime.now()
+        resource.last_modified_internal = datetime.utcnow()
         dataset.update_resource(resource)
-        dataset.last_modified_internal = datetime.now()
+        dataset.last_modified_internal = datetime.utcnow()
         dataset.save()
         return resource
 
@@ -497,7 +499,7 @@ class ResourceAPI(ResourceMixin, API):
         ResourceEditPermission(dataset).test()
         resource = self.get_resource_or_404(dataset, rid)
         dataset.remove_resource(resource)
-        dataset.last_modified_internal = datetime.now()
+        dataset.last_modified_internal = datetime.utcnow()
         dataset.save()
         return '', 204
 
@@ -538,7 +540,7 @@ class CommunityResourcesAPI(API):
             })
         if not resource.organization:
             resource.owner = current_user._get_current_object()
-        resource.last_modified_internal = datetime.now()
+        resource.last_modified_internal = datetime.utcnow()
         resource.save()
         return resource, 201
 
@@ -566,7 +568,7 @@ class CommunityResourceAPI(API):
         form.populate_obj(community)
         if not community.organization and not community.owner:
             community.owner = current_user._get_current_object()
-        community.last_modified_internal = datetime.now()
+        community.last_modified_internal = datetime.utcnow()
         community.save()
         return community
 
@@ -608,14 +610,18 @@ class DatasetSuggestAPI(API):
         '''Datasets suggest endpoint using mongoDB contains'''
         args = suggest_parser.parse_args()
         datasets_query = Dataset.objects(archived=None, deleted=None, private=False)
-        datasets = datasets_query.filter(Q(title__icontains=args['q']) | Q(acronym__icontains=args['q']))
+        datasets = datasets_query.filter(
+            Q(title__icontains=args['q']) | Q(acronym__icontains=args['q']))
         return [
             {
                 'id': dataset.id,
                 'title': dataset.title,
                 'acronym': dataset.acronym,
                 'slug': dataset.slug,
-                'image_url': dataset.organization.logo if dataset.organization else dataset.owner.avatar if dataset.owner else None
+                'image_url': (
+                    dataset.organization.logo if dataset.organization
+                    else dataset.owner.avatar if dataset.owner else None
+                )
             }
             for dataset in datasets.order_by(SUGGEST_SORTING).limit(args['size'])
         ]
@@ -628,7 +634,8 @@ class FormatsSuggestAPI(API):
     def get(self):
         '''Suggest file formats'''
         args = suggest_parser.parse_args()
-        results = [{'text': i} for i in current_app.config['ALLOWED_RESOURCES_EXTENSIONS'] if args['q'] in i]
+        results = [{'text': i} for i in current_app.config['ALLOWED_RESOURCES_EXTENSIONS']
+                   if args['q'] in i]
         results = results[:args['size']]
         return sorted(results, key=lambda o: len(o['text']))
 
@@ -640,7 +647,8 @@ class MimesSuggestAPI(API):
     def get(self):
         '''Suggest mime types'''
         args = suggest_parser.parse_args()
-        results = [{'text': i} for i in current_app.config['ALLOWED_RESOURCES_MIMES'] if args['q'] in i]
+        results = [{'text': i} for i in current_app.config['ALLOWED_RESOURCES_MIMES']
+                   if args['q'] in i]
         results = results[:args['size']]
         return sorted(results, key=lambda o: len(o['text']))