+++ udata/core/organization/rdf.py
@@ -7,7 +7,7 @@ from flask import url_for
 from rdflib import Graph, URIRef, Literal, BNode
 from rdflib.namespace import RDF, RDFS, FOAF
 
-from udata.rdf import DCAT, DCT, DCAT, namespace_manager, paginate_catalog
+from udata.rdf import DCAT, DCT, namespace_manager, paginate_catalog
 
 from udata.core.dataset.rdf import dataset_to_rdf
 from udata.utils import Paginable
@@ -21,7 +21,7 @@ def organization_to_rdf(org, graph=None):
     graph = graph or Graph(namespace_manager=namespace_manager)
     if org.id:
         org_url = endpoint_for('organizations.show_redirect', 'api.organization',
-                                org=org.id, _external=True)
+                               org=org.id, _external=True)
         id = URIRef(org_url)
     else:
         id = BNode()
@@ -42,13 +42,15 @@ def build_org_catalog(org, datasets, format=None):
     catalog = graph.resource(URIRef(org_catalog_url))
     catalog.set(RDF.type, DCAT.Catalog)
     catalog.set(DCT.publisher, organization_to_rdf(org, graph))
+    catalog.set(DCT.title, Literal(f"{org.name}"))
+    catalog.set(DCT.description, Literal(f"{org.name}"))
 
     for dataset in datasets:
         catalog.add(DCAT.dataset, dataset_to_rdf(dataset, graph))
 
     values = {'org': org.id}
-    
+
     if isinstance(datasets, Paginable):
         paginate_catalog(catalog, graph, datasets, format, 'api.organization_rdf_format', **values)
-    
+
     return catalog