+++ udata/harvest/actions.py
@@ -102,7 +102,7 @@ def update_source(ident, data):
 def validate_source(ident, comment=None):
     '''Validate a source for automatic harvesting'''
     source = get_source(ident)
-    source.validation.on = datetime.now()
+    source.validation.on = datetime.utcnow()
     source.validation.comment = comment
     source.validation.state = VALIDATION_ACCEPTED
     if current_user.is_authenticated:
@@ -116,7 +116,7 @@ def validate_source(ident, comment=None):
 def reject_source(ident, comment):
     '''Reject a source for automatic harvesting'''
     source = get_source(ident)
-    source.validation.on = datetime.now()
+    source.validation.on = datetime.utcnow()
     source.validation.comment = comment
     source.validation.state = VALIDATION_REFUSED
     if current_user.is_authenticated:
@@ -128,7 +128,7 @@ def reject_source(ident, comment):
 def delete_source(ident):
     '''Delete an harvest source'''
     source = get_source(ident)
-    source.deleted = datetime.now()
+    source.deleted = datetime.utcnow()
     source.save()
     signals.harvest_source_deleted.send(source)
     return source
@@ -139,7 +139,7 @@ def clean_source(ident):
     source = get_source(ident)
     datasets = Dataset.objects.filter(harvest__source_id=str(source.id))
     for dataset in datasets:
-        dataset.deleted = datetime.now()
+        dataset.deleted = datetime.utcnow()
         dataset.save()
     return len(datasets)
 
@@ -161,7 +161,7 @@ def purge_sources():
 def purge_jobs():
     '''Delete jobs older than retention policy'''
     retention = current_app.config['HARVEST_JOBS_RETENTION_DAYS']
-    expiration = datetime.now() - timedelta(days=retention)
+    expiration = datetime.utcnow() - timedelta(days=retention)
     return HarvestJob.objects(created__lt=expiration).delete()
 
 
@@ -304,7 +304,7 @@ def attach(domain, filename):
             dataset.harvest.domain = domain
             dataset.harvest.remote_id = row['remote']
 
-            dataset.last_modified_internal = datetime.now()
+            dataset.last_modified_internal = datetime.utcnow()
             dataset.save()
             count += 1