+++ udata/tests/api/test_datasets_api.py
@@ -111,13 +111,22 @@ class DatasetAPITest(APITestCase):
 
         second.title = "second updated dataset"
         second.save()
-        response = self.get(url_for('api.datasets', sort='-last_modified'))
+        response = self.get(url_for('api.datasets', sort='-last_update'))
         self.assert200(response)
         self.assertEqual(response.json['data'][0]['id'], str(second.id))
 
-        response = self.get(url_for('api.datasets', sort='last_modified'))
+        response = self.get(url_for('api.datasets', sort='last_update'))
         self.assert200(response)
         self.assertEqual(response.json['data'][0]['id'], str(first.id))
+
+    def test_dataset_api_default_sorting(self):
+        # Default sort should be -created
+        self.login()
+        [VisibleDatasetFactory(title="some created dataset") for i in range(10)]
+        last = VisibleDatasetFactory(title="last created dataset")
+        response = self.get(url_for('api.datasets'))
+        self.assert200(response)
+        self.assertEqual(response.json['data'][0]['id'], str(last.id))
 
     def test_dataset_api_list_with_filters(self):
         '''Should filters datasets results based on query filters'''
@@ -229,7 +238,7 @@ class DatasetAPITest(APITestCase):
 
     def test_dataset_api_get_deleted(self):
         '''It should not fetch a deleted dataset from the API and raise 410'''
-        dataset = VisibleDatasetFactory(deleted=datetime.now())
+        dataset = VisibleDatasetFactory(deleted=datetime.utcnow())
 
         response = self.get(url_for('api.dataset', dataset=dataset))
         self.assert410(response)
@@ -238,7 +247,7 @@ class DatasetAPITest(APITestCase):
         '''It should a deleted dataset from the API if user is authorized'''
         self.login()
         dataset = VisibleDatasetFactory(owner=self.user,
-                                        deleted=datetime.now())
+                                        deleted=datetime.utcnow())
 
         response = self.get(url_for('api.dataset', dataset=dataset))
         self.assert200(response)
@@ -534,7 +543,7 @@ class DatasetAPITest(APITestCase):
     def test_dataset_api_update_deleted(self):
         '''It should not update a deleted dataset from the API and raise 401'''
         user = self.login()
-        dataset = DatasetFactory(owner=user, deleted=datetime.now())
+        dataset = DatasetFactory(owner=user, deleted=datetime.utcnow())
         data = dataset.to_dict()
         data['description'] = 'new description'
         response = self.put(url_for('api.dataset', dataset=dataset), data)
@@ -560,7 +569,7 @@ class DatasetAPITest(APITestCase):
     def test_dataset_api_delete_deleted(self):
         '''It should delete a deleted dataset from the API and raise 410'''
         user = self.login()
-        dataset = VisibleDatasetFactory(owner=user, deleted=datetime.now())
+        dataset = VisibleDatasetFactory(owner=user, deleted=datetime.utcnow())
         response = self.delete(url_for('api.dataset', dataset=dataset))
 
         self.assert410(response)
@@ -1272,7 +1281,7 @@ class DatasetArchivedAPITest(APITestCase):
     def test_dataset_api_search_archived(self):
         '''It should search datasets from the API, excluding archived ones'''
         VisibleDatasetFactory(archived=None)
-        dataset = VisibleDatasetFactory(archived=datetime.now())
+        dataset = VisibleDatasetFactory(archived=datetime.utcnow())
 
         response = self.get(url_for('api.datasets', q=''))
         self.assert200(response)
@@ -1282,7 +1291,7 @@ class DatasetArchivedAPITest(APITestCase):
 
     def test_dataset_api_get_archived(self):
         '''It should fetch an archived dataset from the API and return 200'''
-        dataset = VisibleDatasetFactory(archived=datetime.now())
+        dataset = VisibleDatasetFactory(archived=datetime.utcnow())
         response = self.get(url_for('api.dataset', dataset=dataset))
         self.assert200(response)