+++ udata/tests/api/test_organizations_api.py
@@ -15,11 +15,12 @@ from udata.core.organization.factories import OrganizationFactory
 from udata.core.user.factories import UserFactory, AdminFactory
 from udata.core.dataset.factories import DatasetFactory
 from udata.core.reuse.factories import ReuseFactory
+from udata.i18n import _
 
 from udata.tests.helpers import (
     assert_emit, assert_not_emit,
-    assert200, assert201, assert204, assert403, assert404, assert410,
-    assert_status
+    assert200, assert201, assert204, assert403, assert400, assert404,
+    assert410, assert_status
 )
 
 pytestmark = [
@@ -51,7 +52,7 @@ class OrganizationAPITest:
 
     def test_organization_api_get_deleted(self, api):
         '''It should not fetch a deleted organization from the API'''
-        organization = OrganizationFactory(deleted=datetime.now())
+        organization = OrganizationFactory(deleted=datetime.utcnow())
         response = api.get(url_for('api.organization', org=organization))
         assert410(response)
 
@@ -59,7 +60,7 @@ class OrganizationAPITest:
         '''It should fetch a deleted organization from the API if authorized'''
         user = api.login()
         member = Member(user=user, role='editor')
-        organization = OrganizationFactory(deleted=datetime.now(),
+        organization = OrganizationFactory(deleted=datetime.utcnow(),
                                            members=[member])
         response = api.get(url_for('api.organization', org=organization))
         assert200(response)
@@ -90,9 +91,42 @@ class OrganizationAPITest:
         assert Organization.objects.count() is 1
         assert Organization.objects.first().description == 'new description'
 
+    def test_organization_api_update_business_number_id(self, api):
+        '''It should update an organization from the API by adding a business number id'''
+        user = api.login()
+        member = Member(user=user, role='admin')
+        org = OrganizationFactory(members=[member])
+        data = org.to_dict()
+        data['business_number_id'] = '13002526500013'
+        response = api.put(url_for('api.organization', org=org), data)
+        assert200(response)
+        assert Organization.objects.count() is 1
+        assert Organization.objects.first().business_number_id == '13002526500013'
+
+    def test_organization_api_update_business_number_id_failing(self, api):
+        '''It should update an organization from the API by adding a business number id'''
+        user = api.login()
+        member = Member(user=user, role='admin')
+        org = OrganizationFactory(members=[member])
+        data = org.to_dict()
+        data['business_number_id'] = '110014016'
+        response = api.put(url_for('api.organization', org=org), data)
+        assert400(response)
+        assert response.json['errors']['business_number_id'][0] == _('A siret number is made of 14 digits')
+
+        data['business_number_id'] = '12345678901234'
+        response = api.put(url_for('api.organization', org=org), data)
+        assert400(response)
+        assert response.json['errors']['business_number_id'][0] == _('Invalid Siret number')
+
+        data['business_number_id'] = 'tttttttttttttt'
+        response = api.put(url_for('api.organization', org=org), data)
+        assert400(response)
+        assert response.json['errors']['business_number_id'][0] == _('A siret number is only made of digits')
+
     def test_organization_api_update_deleted(self, api):
         '''It should not update a deleted organization from the API'''
-        org = OrganizationFactory(deleted=datetime.now())
+        org = OrganizationFactory(deleted=datetime.utcnow())
         data = org.to_dict()
         data['description'] = 'new description'
         api.login()
@@ -124,7 +158,7 @@ class OrganizationAPITest:
     def test_organization_api_delete_deleted(self, api):
         '''It should not delete a deleted organization from the API'''
         api.login()
-        organization = OrganizationFactory(deleted=datetime.now())
+        organization = OrganizationFactory(deleted=datetime.utcnow())
         response = api.delete(url_for('api.organization', org=organization))
         assert410(response)
         assert Organization.objects[0].deleted is not None