+++ udata/tests/test_model.py
@@ -200,6 +200,24 @@ class SlugFieldTest:
         assert len(obj.title) == SlugTester.slug.max_length + 1
         assert len(obj.slug) == SlugTester.slug.max_length
 
+    def test_crop_with_index(self):
+        '''SlugField should truncate itself to keep room for index suffix if not unique'''
+        first_obj = SlugTester(title='x' * (SlugTester.slug.max_length + 1))
+        first_obj.save()
+        assert len(first_obj.slug) == SlugTester.slug.max_length
+        assert first_obj.slug.endswith('x')
+        # Try adding a second obj with same title with a slug already at max_length
+        second_obj = SlugTester(title='x' * (SlugTester.slug.max_length + 1))
+        second_obj.save()
+        assert len(second_obj.slug) == SlugTester.slug.max_length
+        assert second_obj.slug.endswith('-1')
+        # We could even have 10+ obj with the same title that needs index suffix
+        [SlugTester(title='x' * (SlugTester.slug.max_length + 1)).save() for i in range(8)]
+        last_obj = SlugTester(title='x' * (SlugTester.slug.max_length + 1))
+        last_obj.save()
+        assert len(last_obj.slug) == SlugTester.slug.max_length
+        assert last_obj.slug.endswith('-10')
+
     def test_multiple_spaces(self):
         field = db.SlugField()
         assert field.slugify('a  b') == 'a-b'
@@ -375,21 +393,21 @@ class DatetimedTest:
         assert isinstance(DatetimedTester.last_modified, db.DateTimeField)
 
     def test_new_instance(self):
-        now = datetime.now()
+        now = datetime.utcnow()
         datetimed = DatetimedTester()
 
-        assert now <= datetimed.created_at <= datetime.now()
-        assert now <= datetimed.last_modified <= datetime.now()
+        assert now <= datetimed.created_at <= datetime.utcnow()
+        assert now <= datetimed.last_modified <= datetime.utcnow()
 
     def test_save_new_instance(self):
-        now = datetime.now()
+        now = datetime.utcnow()
         datetimed = DatetimedTester.objects.create()
 
-        assert now <= datetimed.created_at <= datetime.now()
-        assert now <= datetimed.last_modified <= datetime.now()
+        assert now <= datetimed.created_at <= datetime.utcnow()
+        assert now <= datetimed.last_modified <= datetime.utcnow()
 
     def test_save_last_modified_instance(self):
-        now = datetime.now()
+        now = datetime.utcnow()
         earlier = now - timedelta(days=1)
         datetimed = DatetimedTester.objects.create(
             created_at=earlier, last_modified=earlier)
@@ -401,7 +419,7 @@ class DatetimedTest:
         assert_equal_dates(datetimed.last_modified, now)
 
     def test_save_last_modified_instance_manually_set(self):
-        now = datetime.now()
+        now = datetime.utcnow()
         manual = now - timedelta(days=1)
         earlier = now - timedelta(days=2)
         datetimed = DatetimedTester.objects.create(created_at=earlier, last_modified=earlier)
@@ -414,7 +432,7 @@ class DatetimedTest:
         assert_equal_dates(datetimed.last_modified, manual)
 
     def test_save_last_modified_instance_manually_set_same_value(self):
-        now = datetime.now()
+        now = datetime.utcnow()
         manual = now - timedelta(days=1)
         earlier = now - timedelta(days=2)
         datetimed = DatetimedTester.objects.create(created_at=earlier, last_modified=earlier)
@@ -436,7 +454,7 @@ class ExtrasFieldTest:
         class Tester(db.Document):
             extras = db.ExtrasField()
 
-        now = datetime.now()
+        now = datetime.utcnow()
         today = date.today()
 
         tester = Tester(extras={
@@ -469,7 +487,7 @@ class ExtrasFieldTest:
         (db.IntField, 42),
         (db.FloatField, 0.42),
         (db.BooleanField, True),
-        (db.DateTimeField, datetime.now()),
+        (db.DateTimeField, datetime.utcnow()),
         (db.DateField, date.today()),
     ])
     def test_validate_registered_db_type(self, dbtype, value):
@@ -481,8 +499,8 @@ class ExtrasFieldTest:
         Tester(extras={'test': value}).validate()
 
     @pytest.mark.parametrize('dbtype,value', [
-        (db.IntField, datetime.now()),
-        (db.FloatField, datetime.now()),
+        (db.IntField, datetime.utcnow()),
+        (db.FloatField, datetime.utcnow()),
         (db.BooleanField, 42),
         (db.DateTimeField, 42),
         (db.DateField, 42),